Revert "refactor: move `common` package under `testkit/common`"
