+++ plans/001-big-blocks/tests/app-sync/run_seeds.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/celestiaorg/test-infra/testkit"
 	"github.com/celestiaorg/test-infra/testkit/appkit"
-	"github.com/celestiaorg/test-infra/testkit/common"
+	"github.com/celestiaorg/test-infra/plans/001-big-blocks/tests/common"
 
 	"github.com/testground/sdk-go/network"
 	"github.com/testground/sdk-go/run"
