+++ plans/001-big-blocks/tests/fund-accs/run_validator.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/celestiaorg/test-infra/testkit"
-	"github.com/celestiaorg/test-infra/testkit/common"
+	"github.com/celestiaorg/test-infra/plans/001-big-blocks/tests/common"
 
 	"github.com/testground/sdk-go/network"
 	"github.com/testground/sdk-go/run"
